+++ typescript/cli/src/tests/ethereum/warp/warp-deploy.e2e-test.ts
@@ -850,7 +850,7 @@ describe('hyperlane warp deploy e2e tests', async function () {
           privateKey: HYP_KEY_BY_PROTOCOL.ethereum,
         })
         .should.be.rejectedWith(
-          'Error: Origin (anvil1) or destination (anvil3) are not part of the warp route.',
+          'Error: Chain(s) anvil1 are not part of the warp route.',
         );
     });