+++ typescript/utils/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @hyperlane-xyz/utils
 
+## 19.1.1
+
+## 19.1.0
+
+## 19.0.0
+
+### Major Changes
+
+- 32479e139: feat: add AltVM interface
+
+### Minor Changes
+
+- 8eab305bd: chore: add transactionToPrintableJson to altvm interface
+- e42a0e8e1: chore: updated AltVM interface
+
 ## 18.3.0
 
 ### Minor Changes